move LikeContentButton to be child of LikeContentButtonConnected in ui-connected
