+++ packages/apollos-ui-connected/src/LikeContentButtonConnected/LikeContentButton/index.js
