+++ packages/apollos-ui-connected/src/LikeContentButtonConnected/index.js
@@ -5,8 +5,8 @@ import { get } from 'lodash';
 import { Query, Mutation } from 'react-apollo';
 
 import { AnalyticsConsumer } from '@apollosproject/ui-analytics';
-import LikeContentButton from '../LikeContentButton';
 
+import LikeContentButton from './LikeContentButton';
 import UPDATE_LIKE_ENTITY from './updateLikeEntity';
 import GET_LIKED_CONTENT_ITEM from './getLikedContentItem';
 import updateLikedContent from './updateLikedContent';
